test: split the Identity database constraint cases and cover removed-key ordering
