+++ charts/camunda-platform-8.10/test/unit/common/constraints_test.go
@@ -909,165 +909,3 @@ func (s *ConstraintTemplateTest) TestManagementIdentityExternalServiceUrl() {
 
 	testhelpers.RunTestCasesE(s.T(), s.chartPath, s.release, s.namespace, s.templates, testCases)
 }
-
-func (s *ConstraintTemplateTest) TestIdentityDatabaseRequiredForNonKeycloakAuth() {
-	identityOidcValues := func(extra map[string]string) map[string]string {
-		values := map[string]string{
-			"orchestration.data.secondaryStorage.type": "elasticsearch",
-			"identity.enabled":                         "true",
-			"global.identity.auth.enabled":             "true",
-			"global.identity.auth.type":                "GENERIC",
-			"global.identity.auth.issuer":              "https://issuer.example.com",
-			"global.identity.auth.issuerBackendUrl":    "https://issuer.example.com",
-			"global.identity.auth.tokenUrl":            "https://issuer.example.com/token",
-			"global.identity.auth.jwksUrl":             "https://issuer.example.com/keys",
-			"global.identity.auth.authUrl":             "https://issuer.example.com/auth",
-		}
-		for key, value := range extra {
-			values[key] = value
-		}
-		return values
-	}
-
-	testCases := []testhelpers.TestCase{
-		{
-			Name:   "GenericAuthWithoutDatabaseFails",
-			Values: identityOidcValues(nil),
-			Verifier: func(t *testing.T, output string, err error) {
-				s.Require().ErrorContains(err, "Management Identity requires a database")
-			},
-		},
-		{
-			Name: "MicrosoftAuthWithoutDatabaseFails",
-			Values: identityOidcValues(map[string]string{
-				"global.identity.auth.type": "MICROSOFT",
-			}),
-			Verifier: func(t *testing.T, output string, err error) {
-				s.Require().ErrorContains(err, "Management Identity requires a database")
-			},
-		},
-		{
-			Name: "GenericAuthWithExternalDatabaseRenders",
-			Values: identityOidcValues(map[string]string{
-				"identity.externalDatabase.enabled": "true",
-			}),
-			Verifier: func(t *testing.T, output string, err error) {
-				s.Require().NoError(err)
-			},
-		},
-		{
-			Name: "GenericAuthWithDatabaseEnvRenders",
-			Values: identityOidcValues(map[string]string{
-				"identity.env[0].name":  "IDENTITY_DATABASE_HOST",
-				"identity.env[0].value": "postgres.example.com",
-			}),
-			Verifier: func(t *testing.T, output string, err error) {
-				s.Require().NoError(err)
-			},
-		},
-		{
-			Name: "GenericAuthWithUnrelatedEnvFails",
-			Values: identityOidcValues(map[string]string{
-				"identity.env[0].name":  "TZ",
-				"identity.env[0].value": "UTC",
-			}),
-			Verifier: func(t *testing.T, output string, err error) {
-				s.Require().ErrorContains(err, "Management Identity requires a database")
-			},
-		},
-		{
-			Name: "GenericAuthWithEnvFromRenders",
-			Values: identityOidcValues(map[string]string{
-				"identity.envFrom[0].secretRef.name": "identity-database",
-			}),
-			Verifier: func(t *testing.T, output string, err error) {
-				s.Require().NoError(err)
-			},
-		},
-		{
-			Name: "GenericAuthWithDatasourceInConfigurationRenders",
-			Values: identityOidcValues(map[string]string{
-				"identity.configuration": "spring:\n  datasource:\n    url: jdbc:postgresql://postgres.example.com/identity",
-			}),
-			Verifier: func(t *testing.T, output string, err error) {
-				s.Require().NoError(err)
-			},
-		},
-		{
-			Name: "GenericAuthWithDatasourceInExtraConfigurationRenders",
-			Values: identityOidcValues(map[string]string{
-				"identity.extraConfiguration[0].file":    "datasource.yaml",
-				"identity.extraConfiguration[0].content": "spring:\n  datasource:\n    url: jdbc:postgresql://postgres.example.com/identity",
-			}),
-			Verifier: func(t *testing.T, output string, err error) {
-				s.Require().NoError(err)
-			},
-		},
-		{
-			Name: "KeycloakAuthWithoutDatabaseRenders",
-			Values: map[string]string{
-				"orchestration.data.secondaryStorage.type": "elasticsearch",
-				"identity.enabled":                         "true",
-				"global.identity.auth.enabled":             "true",
-			},
-			Verifier: func(t *testing.T, output string, err error) {
-				s.Require().NoError(err)
-			},
-		},
-		{
-			Name: "GenericAuthWithIdentityDisabledRenders",
-			Values: identityOidcValues(map[string]string{
-				"identity.enabled": "false",
-			}),
-			Verifier: func(t *testing.T, output string, err error) {
-				s.Require().NoError(err)
-			},
-		},
-		{
-			Name: "GenericAuthWithExternalManagementIdentityRenders",
-			Values: identityOidcValues(map[string]string{
-				"identity.enabled":            "false",
-				"global.identity.service.url": "http://identity.other-ns.svc:8080",
-			}),
-			Verifier: func(t *testing.T, output string, err error) {
-				s.Require().NoError(err)
-			},
-		},
-		{
-			Name: "MultiTenancyWithoutIdentityDatabaseFails",
-			Values: map[string]string{
-				"orchestration.data.secondaryStorage.type": "elasticsearch",
-				"identity.enabled":                         "true",
-				"global.identity.auth.enabled":             "true",
-				"global.multitenancy.enabled":              "true",
-			},
-			Verifier: func(t *testing.T, output string, err error) {
-				s.Require().ErrorContains(err, "Multi-Tenancy feature")
-			},
-		},
-		{
-			Name: "MultiTenancyWithIdentityDatabaseRenders",
-			Values: map[string]string{
-				"orchestration.data.secondaryStorage.type": "elasticsearch",
-				"identity.enabled":                         "true",
-				"global.identity.auth.enabled":             "true",
-				"global.multitenancy.enabled":              "true",
-				"identity.externalDatabase.enabled":        "true",
-			},
-			Verifier: func(t *testing.T, output string, err error) {
-				s.Require().NoError(err)
-			},
-		},
-		{
-			Name: "GenericAuthWithIdentityAuthDisabledRenders",
-			Values: identityOidcValues(map[string]string{
-				"global.identity.auth.enabled": "false",
-			}),
-			Verifier: func(t *testing.T, output string, err error) {
-				s.Require().NoError(err)
-			},
-		},
-	}
-
-	testhelpers.RunTestCasesE(s.T(), s.chartPath, s.release, s.namespace, s.templates, testCases)
-}
